Add test data
The scraper is not currently capturing data for all members. It is also pulling in malformed data. For example, Arben Imami's birth place is recorded as `21.01.1958 Vendlindja: Tiranë`.

The fixtures are being added to reflect the current state of the scraper. The purpose of this PR is solely to extract the MemberPage class. Changes to the data output will be made in a forthcoming PR.